Move TTSManager around.
