+++ Health/src/main/java/com/sciencesquad/health/core/util/TTSManager.java
@@ -1,4 +1,4 @@
-package com.sciencesquad.health.run;
+package com.sciencesquad.health.core.util;
 
 import android.content.Context;
 import android.speech.tts.TextToSpeech;
@@ -12,27 +12,25 @@ public class TTSManager {
 
     public void init(Context context) {
         try {
-            mTts = new TextToSpeech(context, onInitListener);
+            mTts = new TextToSpeech(context, new TextToSpeech.OnInitListener() {
+                @Override
+                public void onInit(int status) {
+                    if (status == TextToSpeech.SUCCESS) {
+                        int result = mTts.setLanguage(Locale.US);
+                        isLoaded = true;
+
+                        if (result == TextToSpeech.LANG_MISSING_DATA || result == TextToSpeech.LANG_NOT_SUPPORTED) {
+                            Log.e("error", "This Language is not supported");
+                        }
+                    } else {
+                        Log.e("error", "Initialization Failed!");
+                    }
+                }
+            });
         } catch (Exception e) {
             e.printStackTrace();
         }
     }
-
-    private TextToSpeech.OnInitListener onInitListener = new TextToSpeech.OnInitListener() {
-        @Override
-        public void onInit(int status) {
-            if (status == TextToSpeech.SUCCESS) {
-                int result = mTts.setLanguage(Locale.US);
-                isLoaded = true;
-
-                if (result == TextToSpeech.LANG_MISSING_DATA || result == TextToSpeech.LANG_NOT_SUPPORTED) {
-                    Log.e("error", "This Language is not supported");
-                }
-            } else {
-                Log.e("error", "Initialization Failed!");
-            }
-        }
-    };
 
     public void addQueue(String text) {
         if (isLoaded) {
